requirements.txt
accelerate==0.22.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
appdirs==1.4.4
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1677499911949/work
brotlipy==0.7.0
certifi==2023.7.22
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chardet @ file:///home/builder/ci_310/chardet_1640804867535/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cmaes==0.10.0
colorama==0.4.6
colorlog==6.7.0
Columnar==1.4.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1692727121011/work
conda-build==3.23.3
conda-package-handling @ file:///croot/conda-package-handling_1666940373510/work
contourpy==1.1.0
cosine-annealing-warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
cryptography @ file:///croot/cryptography_1665612644927/work
cupy-cuda11x==12.2.0
cycler==0.11.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dnspython==2.2.1
docker-pycreds==0.4.0
einops==0.6.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup==1.0.4
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
expecttest==0.1.4
fastrlock==0.8.2
filelock @ file:///opt/conda/conda-bld/filelock_1647002191454/work
finance-datareader==0.9.66
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.42.1
frozendict==2.3.10
frozenlist==1.4.0
fsspec==2023.9.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.32
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
greenlet==2.0.2
h5py==3.9.0
html5lib==1.1
huggingface-hub==0.17.2
hypothesis==6.61.0
idna @ file:///croot/idna_1666125576474/work
ipykernel==6.28.0
ipython @ file:///croot/ipython_1670919316550/work
ipywidgets==8.1.1
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib==1.3.2
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1632759296524/work
jsonpointer==2.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1669775088561/work
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
lightning-utilities==0.9.0
llvmlite==0.41.0
lxml==4.9.3
Mako==1.2.4
MarkupSafe @ file:///home/builder/ci_310/markupsafe_1640791577099/work
matplotlib==3.7.2
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
mpmath==1.2.1
msgpack==1.0.7
multidict==6.0.4
multitasking==0.0.11
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numba==0.58.0
numpy==1.25.2
openTSNE==1.0.0
optuna==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==1.5.3
pandas-datareader==0.10.0
pandas-parallel-apply==2.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pastel==0.2.1
pathtools==0.1.2
patsy==0.5.6
peewee==3.17.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
pkginfo @ file:///croot/pkginfo_1666725041340/work
plotly==5.18.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
poethepoet==0.10.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==4.24.1
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-etcd==0.4.5
pytorch-lightning==2.1.0
pytz==2023.3.post1
PyYAML==6.0
pyzmq @ file:///croot/pyzmq_1686601365461/work
ray==2.7.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
rpds-py==0.10.6
rtdl_num_embeddings==0.0.9
rtdl_revisiting_models==0.0.2
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
sentry-sdk==1.29.2
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve @ file:///croot/soupsieve_1666296392845/work
SQLAlchemy==2.0.21
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
statsmodels==0.14.1
string-color==1.2.3
sympy==1.11.1
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomlkit==0.12.2
toolz @ file:///croot/toolz_1667464077321/work
torch==1.13.1
torchcde==0.2.5
torchdiffeq==0.2.3
torchdyn==1.0.6
torchelastic==0.2.2
torchmetrics==1.2.0
torchpq==0.3.0.1
torchsde==0.2.6
torchtext==0.14.1
torchvision==0.14.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827254365/work
tqdm==4.66.1
traitlets @ file:///croot/traitlets_1671143879854/work
trampoline==0.1.2
transformers==4.33.2
types-dataclasses==0.6.6
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3 @ file:///croot/urllib3_1670526988650/work
-e git+https://github.com/minyoungg/vqtorch@d0022eabae6c5628bf7297cffc3e0ba46cf10392#egg=vqtorch
wandb==0.16.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
widgetsnbextension==4.0.9
x-transformers==1.21.3
yarl==1.9.2
yfinance==0.2.32
